vaargs from ofutils to oflog
